Replace direct 'ssh proxy' execution with crypto/ssh call, like what was already done with tls routing enabled.
